Fix offset type
git-svn-id: https://legesmotus.svn.sourceforge.net/svnroot/legesmotus/legesmotus/trunk@182 a5a5d5d7-22d1-44d1-b37a-67d3d99fcfd7

+++ client/GameWindow.hpp
@@ -23,8 +23,8 @@ private:
 	std::list<Graphic*>		m_graphics;
 	std::list<Graphic*>		m_hud_graphics;
 
-	int			m_offset_x;
-	int			m_offset_y;
+	double			m_offset_x;
+	double			m_offset_y;
 
 	GameWindow(int width, int height, int depth, bool fullscreen);
 	~GameWindow();
